+++ docs/CHANGELOG.md
@@ -12,6 +12,8 @@ Known [issues](https://github.com/vengi-voxel/vengi/issues?q=is%3Aissue+is%3Aope
 
 General:
 
+   - Fixed an off-by-one bug in the marching cubes extractor
+
 ## 0.0.37 (2025-05-12)
 
 General: